+++ citizen-intelligence-agency/.vscode/settings.json
@@ -4,8 +4,7 @@
     "java.dependency.packagePresentation": "hierarchical",
     "maven.workspace.autoReload": true,
 
-    "maven.workspace.resolveFolders": false, // Disable workspace resolution
-    "maven.terminal.defaultProfiles": ["use-m2"], // Custom Maven profile to enforce .m2 resolution
+    "maven.workspace.resolveFolders": true, // Disable workspace resolution
   
     // Recommended Maven commands
     "maven.terminal.customGoals": ["clean", "install", "verify"],
@@ -16,16 +15,16 @@
       "-Denforcer.skip=true"
     ],
     "files.exclude": {
-        "**/module-info.java": true,
-        "**/module-info.class": true
-      },
-      "files.watcherExclude": {
-        "**/module-info.java": true,
-        "**/module-info.class": true
-      },
-      "java.import.exclusions": [
-        "**/module-info.java",
-        "**/module-info.class"
-      ],
-      "maven.view": "hierarchical"
+      "**/module-info.java": true,
+      "**/module-info.class": true
+    },
+    "files.watcherExclude": {
+      "**/module-info.java": true,
+      "**/module-info.class": true
+    },
+    "java.import.exclusions": [
+      "**/module-info.java",
+      "**/module-info.class"
+    ],
+    "maven.view": "hierarchical"
 }
