+++ parent-model-internal-pom/pom.xml
@@ -188,8 +188,8 @@
         <groupId>org.bsc.maven</groupId>
         <artifactId>maven-processor-plugin</artifactId>
         <inherited>true</inherited>
-        <?m2e execute onConfiguration?>	
-        <?m2e execute onIncremental?>	
+        <?m2e execute onConfiguration?>
+        <?m2e execute onIncremental?>
         <executions>
           <execution>
             <id>process</id>
